System card now working.
System card variables with parentheis (indexes) now work by submitting a list or
a tuple of length 2, where the first value is the index and the second value is
the parameter to be used. Note, will have to check starting_ns_eigenvalue to see
how it needs to be handled.