Added local and queue run functionality. Changed to jasp style of execute. Added examples.
Now can run local or sub to queue with calculate(mode="local" or "queue"). Can
add parallelization flags. Added si_sub.py example for submission. Changed to
jasp style, syntax: with pwx(...) as calc:
try:
  calc.calculate()
except(QepyException):
This is a safe style of execution as the object only exists within the with block.