Changed local run to use Popen. Changed qepyrc read_configuration. Added directory check for pwx.
The mode=local now uses Popen with the .wait() command to force the program to wait until termination.
Also automatically uses mpirun now if more than one processor is requested. The qepyrc read_configuration
 now splits lines using a simple equals sign and then assigns value, key pairs in the run_keys. The pwx
object now checks to see if it is currently in the requested directory. Robustness of directory checking
upgraded for pwx, phx, and q2r