Added a feature to retrieve forces, stress and pressure
the get_forces function parses the output text file for the force matrix. It then
returns the forces as a numpy array

the get_stess_pressure function parses the output text file for the stress and pressure.
The stress is returned as a numpy array. The pressure is returned as a numpy array for
the pressure matrix, and a separate array for the total pressure. In order to only retrieve
the stress or the pressure pass the parameter type='stress' or 'pressure'